photo_export/requirements.txt
requests
google-auth
google-auth-oauthlib
Pillow
pytest
